Move file one level up
